+++ .gitignore
@@ -1,3 +1,4 @@
+.idea
 target
 Cargo.lock
 **/app_id
