+++ Cargo.toml
@@ -1,5 +1,11 @@
 [workspace]
-members = ["data-resource", "fs-atomic-versions", "fs-index", "fs-utils"]
+members = [
+    "data-resource",
+    "fs-atomic-versions",
+    "fs-index",
+    "fs-utils"
+]
+
 default-members = [
     "data-resource",
     "fs-atomic-versions",
