+++ data-resource/benches/compute_bytes_benchmark.rs
@@ -4,7 +4,11 @@ use pprof::criterion::{Output, PProfProfiler};
 use rand::prelude::*;
 use std::fs;
 
-const FILE_PATHS: [&str; 2] = ["../testdata/lena.jpg", "../testdata/test.pdf"]; // Add files to benchmark here
+const FILE_PATHS: [&str; 2] = [
+    // Add files to benchmark here
+    "../test-assets/lena.jpg",
+    "../test-assets/test.pdf",
+];
 
 fn generate_random_data(size: usize) -> Vec<u8> {
     let mut rng = rand::thread_rng();
